# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CategoryScore:
    properties:
      earned_marks:
        type: number
      max_marks:
        type: number
      percentage:
        type: number
    type: object
...
  models.CreateResourceInput:
    properties:
      batch_short_id:
        example: use GET /batches to pick a real short_id, or omit for global visibility
        type: string
      description:
        example: Quick reference for week 1.
        type: string
      expires_at:
        example: "2026-12-31T23:59:00Z"
        type: string
      is_downloadable:
        example: true
        type: boolean
      module_short_id:
        example: use GET /modules to pick a real short_id
        type: string
      resource_type:
        enum:
        - pdf
        - document
        - video
        - link
        - image
        - code
        - other
        example: pdf
        type: string
      session_short_id:
        example: use GET /sessions to pick a real short_id
        type: string
      title:
        example: Python Cheatsheet — Variables & Data Types
        type: string
      url:
        example: https://cdn.example.com/cheatsheet.pdf
        type: string
    required:
    - resource_type
    - title
    - url
    type: object
...
  models.Resource:
    properties:
      batch_number:
        type: string
      batch_short_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      created_by_name:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      expires_at:
        type: string
      id:
        type: string
      is_downloadable:
        type: boolean
      module_name:
        type: string
      module_short_id:
        type: string
      resource_type:
        description: pdf | document | video | link | image | code | other
        type: string
      session_name:
        type: string
      session_short_id:
        type: string
      short_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  models.StudentScoreBreakdown:
    properties:
      assignments:
        $ref: '#/definitions/models.CategoryScore'
      batch_id:
        type: string
      batch_number:
        type: string
      batch_short_id:
        type: string
      exams:
        $ref: '#/definitions/models.CategoryScore'
      final_score:
        type: number
      projects:
        $ref: '#/definitions/models.CategoryScore'
      rank:
        type: integer
      student_id:
        type: string
      student_name:
        type: string
    type: object
...
  models.UpdateResourceInput:
    properties:
      batch_short_id:
        example: ""
        type: string
      description:
        example: Updated description
        type: string
      expires_at:
        example: "2027-01-01T00:00:00Z"
        type: string
      is_downloadable:
        example: false
        type: boolean
      module_short_id:
        example: ""
        type: string
      resource_type:
        enum:
        - pdf
        - document
        - video
        - link
        - image
        - code
        - other
        example: video
        type: string
      session_short_id:
        example: ""
        type: string
      title:
        example: Updated title
        type: string
      url:
        example: https://cdn.example.com/new-file.pdf
        type: string
    type: object
  models.UpdateScoreWeightsInput:
    properties:
      assignments_weight:
        example: 40
        maximum: 100
        minimum: 0
        type: integer
      exams_weight:
        example: 40
        maximum: 100
        minimum: 0
        type: integer
      projects_weight:
        example: 20
        maximum: 100
        minimum: 0
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: umangagarwal9099@gmail.com
    name: Umang Agarwal
  description: 'Vedex backend API. Authenticate via POST /api/v1/auth/login to get
    a JWT, then pass it as `Authorization: Bearer <token>` on protected endpoints.'
  title: Vedex API
  version: "1.0"
paths:
...
  /batches/{short_id}/leaderboard:
    get:
      description: Computes every enrolled student's weighted final score (assignments
        + exams + projects, per the batch's configured weights) and rank, persists
        it, and returns the sorted leaderboard.
      parameters:
      - description: Batch short ID
        in: path
        name: short_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentScoreBreakdown'
            type: array
        "404":
          description: Batch not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a batch's leaderboard
      tags:
      - scores
...
  /batches/{short_id}/score-weights:
    patch:
      consumes:
      - application/json
      description: Sets how much each category (assignments/exams/projects) contributes
        to the batch's final score and leaderboard ranking. The three weights must
        sum to 100. Restricted to super_admin / team_lead.
      parameters:
      - description: Batch short ID
        in: path
        name: short_id
        required: true
        type: string
      - description: Category weights (must sum to 100)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateScoreWeightsInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error or weights don't sum to 100
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Batch not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set a batch's score weights
      tags:
      - scores
...
  /resources:
    get:
      description: Returns resources scoped to the caller's role — students see global
        resources plus those for their enrolled batches (expired resources excluded);
        mentors see global resources plus those for batches they manage; team_lead/super_admin
        see everything. Supports optional filtering by batch_short_id and resource_type.
      parameters:
      - description: Filter by batch short ID
        in: query
        name: batch_short_id
        type: string
      - description: 'Filter by type: pdf | document | video | link | image | code
          | other'
        in: query
        name: resource_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Resource'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Publish a new learning resource. Upload a file first via POST /upload/resource-file
        and pass the returned URL, or pass an external link directly. Leave batch_short_id
        empty for global visibility (all students), or set it to scope to one batch.
        Restricted to super_admin / team_lead / mentor.
      parameters:
      - description: Resource details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateResourceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Resource'
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create resource
      tags:
      - resources
  /resources/{short_id}:
    delete:
      description: Soft-deletes a resource by its short ID. Restricted to super_admin
        / team_lead / mentor.
      parameters:
      - description: Resource short ID
        in: path
        name: short_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Resource not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete resource
      tags:
      - resources
    get:
      description: Returns a single resource by its short_id.
      parameters:
      - description: Resource short ID
        in: path
        name: short_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resource'
        "404":
          description: Resource not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get resource
      tags:
      - resources
    patch:
      consumes:
      - application/json
      description: Partially update a resource. All fields are optional. Send an empty
        string for batch_short_id/module_short_id/session_short_id to clear that scoping.
        Restricted to super_admin / team_lead / mentor.
      parameters:
      - description: Resource short ID
        in: path
        name: short_id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateResourceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resource'
        "400":
          description: Validation error or no fields provided
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Resource not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update resource
      tags:
      - resources
...
  /users/{id}/score-breakdown:
    get:
      description: Returns one student's score breakdown (assignments/exams/projects/final
        score) per batch, across every batch they've ever been added to.
      parameters:
      - description: Student user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentScoreBreakdown'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a student's score breakdown
      tags:
      - scores
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"